ckit: differentiate between read and write hash lookups
"read" hash lookups should be possible on nodes that are shutting down
and still handing over work. "write" hash lookups should only be
possible on nodes which are active participants.

As part of this change, the list of valid states and state transitions
has changed dramatically. This list of states are:

* StateViewer (view-only, never used for hashing)
* StateParticipant (read/write capable)
* StateTerminating (read only)

And the valid state transitions has been simplified:

* StateViewer -> StateParticipant
* StateParticipant -> StateTerminating

Nodes intended for view-only access should never transition out of that
state; StateTerminating is dedicated to nodes that were one
participating in the cluster but are now shutting down.

To support this change, chash has been made internal, and the hash
algorithms it exposed are now directly exposed by the ckit package
through a Peer-aware Hasher interface.