hash: move hashing-related functionality to a dedicated package
This commit removes all knowledge of consistent hashing from the
high-level ckit package. The new hash package contains the high-level
API for hashing against a cluster, while the previous chash package
remains internal as a low-level API.

Hashers must be registered as an observer to a Node to be automatically
updated when the cluster changes. This will cause Hashers to be out of
date depending on how slow the other registered observers are.