ckit: optionally sychronize state with a hash.Hasher
This commit undos the side-effect introduced by
b295c9207b0c7020b04fe10c8aa3e95313086896, while still allowing hashing
to exist in its own package.

ckit.Config can be given an optional hash.Hasher which is only used for
sychronizing the current set of peers to. This synchronization happens
in the foreground before observers are updated, ensuring that the hasher
will always reflect the current state of the cluster.