ckit: fix deadlock between handling user message and queueing a broadcast
While c1fb68e00532b2d11fd33f7d380a133dc0d725d3 made gossip convergence
happen faster, it introduced the possibility for a deadlock to occur
while handling user messages.

The gossip transmit queue used the NumMembers function of the
memberlist, which grabs a lock on the node state. If UserMsg was called
while the memberlist's node state was locked, we'd deadlock.

I was unable to find the specific code path that caused the deadlock,
but it was at least obvious that c1fb68e introduced it. This commit
changes the transmit queue to load from an atomic field instead, which
is also how memberlist is internally implemented.