ckit: remove name conflict handing on Start
The name conflict handling on Start was prone to errors; if a name
conflict was found, the memberlist was permanently terminated,
preventing the node from gossiping any new messages.

The original intent for name conflict handling was to prevent two nodes
with the same name from ever existing in the cluster. However, due to
the eventual consistency model clustering uses, the existing code does
not succesfully protect against that in all situations.

Given the code didn't work, and caused surprising issues with
terminating the memberlist, it is being removed until better name
conflict handling can be implemented.