resolve issue with 64b alignment in armv7hf
This is primarily intended as a fix
for Alloy, panic-ing on startup on an armv7f
platform (see https://github.com/grafana/alloy/issues/1099).
This is a known bug with Go (https://pkg.go.dev/sync/atomic#pkg-note-BUG).
Using the the primitive atomic functions should
aid in general when running on certain 32b platforms.