fix: refute state collisions from previous node instance
In very unlucky situations, a node may broadcast a state change message
with the same lamport time as a previous instance of that node, except
with a different state.

This situation should be incredibly rare, depending on which nodes are
joined and whether those nodes have fully converged yet.

However, it's still possible, and there was no logic for a node to
refute receiving a message about itself that does not align with its own
view.

This commit detects this scenario, and broadcasts a new state message
upon receiving a conflicting state.