chore: remove redundant lamport increases for local state change
Passing local state changes as if it were a remote message caused the
lamport time to increase twice: once for the observe, and a second time
to "refute a stale message" for each state change.

This isn't necessary and just made it harder to track lamport time
changes when debugging a test.