fix: delete node states upon node departure
A race condition sometimes prevented a node from being able to quickly
rejoin in the viewer state if it previously left in a non-viewer state.

For example:

1. node-a starts the cluster.
2. node-b joins the cluster, eventually transitions to participant, and
   dies.
3. node-b rejoins the cluster, but stays in the viewer state.

node-a caches the participant state of the old node-b instance (with the
higher lamport time), and ignores the zero-time lamport from the new
instance of node-b.

Meanwhile, node-a doesn't report that cached state to node-b during the
first push/pull because node-b isn't in its list of peers anymore (and
won't return until after the push/pull).

The fix for this is to clear cached states immediately after a node
leaves the cluster rather than after push/pull cycles.