misc: improvements to TCP usage, state handling, and convergence time  (#85)
* fix: refute state collisions from previous node instance

In very unlucky situations, a node may broadcast a state change message
with the same lamport time as a previous instance of that node, except
with a different state.

This situation should be incredibly rare, depending on which nodes are
joined and whether those nodes have fully converged yet.

However, it's still possible, and there was no logic for a node to
refute receiving a message about itself that does not align with its own
view.

This commit detects this scenario, and broadcasts a new state message
upon receiving a conflicting state.

* chore: remove redundant lamport increases for local state change

Passing local state changes as if it were a remote message caused the
lamport time to increase twice: once for the observe, and a second time
to "refute a stale message" for each state change.

This isn't necessary and just made it harder to track lamport time
changes when debugging a test.

* chore: improve TCP settings

The existing settings were tailored more for UDP payloads. Due to packet
writes being handled by a single thread, more generous timeouts help
avoid temporary increases to suspicion when a healthy node is briefly
blocked talking to a dead node.

The specific settings here were copied from the default settings of
grafana/dskit's TCP transport.

* fix: delete node states upon node departure

A race condition sometimes prevented a node from being able to quickly
rejoin in the viewer state if it previously left in a non-viewer state.

For example:

1. node-a starts the cluster.
2. node-b joins the cluster, eventually transitions to participant, and
   dies.
3. node-b rejoins the cluster, but stays in the viewer state.

node-a caches the participant state of the old node-b instance (with the
higher lamport time), and ignores the zero-time lamport from the new
instance of node-b.

Meanwhile, node-a doesn't report that cached state to node-b during the
first push/pull because node-b isn't in its list of peers anymore (and
won't return until after the push/pull).

The fix for this is to clear cached states immediately after a node
leaves the cluster rather than after push/pull cycles.

* misc: add warning to ChangeState

Because the lamport clock is set to zero before joining the cluster, a
node which rejoins the cluster may require to broadcast its state change
twice:

* Once for when it initially joins, and
* once again after joining to fix any conflicts.

This can be avoided by not changing state out of Viewer until after the
node has joined.